Added confirm message before deck deletion

+++ src/app/deck-list/deck-list.component.ts
@@ -1,4 +1,6 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material';
+import { ConfirmComponent } from '../utils/confirm/confirm.component';
 import { DeckService } from '../services/deck.service';
 import { Router } from '@angular/router';
 import { Deck } from '../models/deck';
@@ -14,6 +16,7 @@ export class DeckListComponent implements OnInit {
   selectedDeck: Deck;
   showAdd: boolean = false;
   constructor(
+    public dialog: MatDialog,
     private router: Router,
     private deckService: DeckService) { }
 
@@ -42,11 +45,26 @@ export class DeckListComponent implements OnInit {
   }
 
   deleteDeck(deck:Deck): void {
-    let index = this.decks.indexOf(deck);
-    this.deckService.deleteDeck(deck).subscribe((_) => {
-      this.decks.splice(index, 1);
-      if (this.decks.length == 0) {
-        this.showAdd = true;
+
+    let dialogConfirmRef = this.dialog.open(ConfirmComponent, {
+      width: '600px',
+      data : {
+        title:       'Suppression du deck',
+        message:     'Vous êtes sur le point de supprimer ce deck et toutes ses données.',
+        acceptation: 'Valider',
+        annulation:  'Annuler'
+      },
+      disableClose: true
+    });
+    dialogConfirmRef.afterClosed().subscribe((data:any) => {
+      if (data.result) {
+        let index = this.decks.indexOf(deck);
+        this.deckService.deleteDeck(deck).subscribe((_) => {
+          this.decks.splice(index, 1);
+          if (this.decks.length == 0) {
+            this.showAdd = true;
+          }
+        });
       }
     });
   }
